Make sure that block/widget search result is a number

+++ include/classes.php
@@ -738,11 +738,11 @@ class mf_bank_id
 		$meta_value = check_var($meta_key);
 		$meta_text = __("Social Security Number", 'lang_bank_id');
 
-		$post_id = apply_filters('get_block_search', 'mf/custom_registration');
+		$post_id = (int)apply_filters('get_block_search', 'mf/custom_registration');
 
 		if(!($post_id > 0))
 		{
-			$post_id = apply_filters('get_widget_search', 'registration-widget');
+			$post_id = (int)apply_filters('get_widget_search', 'registration-widget');
 		}
 
 		if($post_id > 0)
